feat: move geolocate helper to sre module
