+++ app/modules/sre/geolocate_helper.py
@@ -2,7 +2,6 @@
 
 This module provides the ability to geolocate an IP address using the maxmind service.
 """
-
 from integrations import maxmind
